* etc/styles/LOAllFactoryStyles.odt: New
* etc/styles/LOAllFactoryStyles.odt: New

This file is created by running `CreateODTFileWithAllFactoryStyles'
function in `contrib/odt/LibreOffice/src/OrgModeUtilities.bas'.

The ODT file is created on the following version of LibreOffice on Nov
19. 2024.

    Version                 : 24.8.2.1 (X86_64) / LibreOffice Community
    Build ID                : 480(Build:1)
    Locale                  : en-IN (en_IN); UI     : en-US
    Debian package version  : 4:24.8.2-2

See earlier commit ea68e825af210f1726fa03e9fddafedf67a61894.

The following is an extract of the above commit:

* OrgModeUtilities.bas (CreateODTFileWithAllFactoryStyles): Create a ODT
file with all vanilla LibreOffice styles.  In a single shot, one can
see the XML representation of all factory styles that LibreOffice
ships with.  The styles that are of specific interest to the current
state of the ODT exporter are: (a) the modern List Styles (b) the
table template definition, and its related styles.

`CreateODTFileWithAllFactoryStyles' is related to #251 (Upgrading
of ODT exporter's default `styles.xml' to the current state of art
XML definitions)

With this change, one can accomplish the following:

- Deprecate or re-define current ODT deployed list styles
  `OrgNumberedList', `OrgBulletedList'--these are "legacy" list
  definitions--to to modern list styles.

  This in turn /may/ allow one fine-grained control over how lists are
  indented. See #56 and #79.

- New LO-supplied styles `List 1', `List 1 start', `List 1 Cont' and
  `List 1 End' etc can be used as starting points for paragraph styles
  that can be used within lists.  See #250.

- Import fancy table cell definitions used in autoformatting of
  Tables.  See #199.

See also the following https://ask.libreoffice.org post

    How do I create a `ott’ template that includes “default” style
    definitions; by “default” styles, I mean ALL the “factory styles”
    that ship with a LibreOffice install:
    https://ask.libreoffice.org/t/how-do-i-create-a-ott-template-that-includes-default-style-definitions-by-default-styles-i-mean-all-the-factory-styles-that-ship-with-a-libreoffice-install/96312